Compute per-row bases in int64 in the fused indexer kernel
tl.program_id is int32 and Triton does not promote integer multiplies,
so the row base pid_b * s_out_b overflows once B * max_seq_len crosses
2^31 -- reachable at 1M-context prefill shapes (a 4096-row call crosses
at msl 524289). Bisected on hardware: the first step past 2^31 raises an
illegal memory access, one step below is clean. Promote pid_b to int64
at the top so every per-row base (output, page table, weights, query)
inherits the wide type; the KV-pool side already carried its own int64
cast. Found by auditing against the same bug class reported in the
vLLM-side indexer fallback (vllm-project/vllm#50576).

+++ python/sglang/kernels/ops/attention/dsv4/fused_paged_indexer.py
@@ -73,7 +73,12 @@ def _fused_paged_indexer_kernel(
     H: tl.constexpr,
     BN: tl.constexpr,
 ):
-    pid_b = tl.program_id(0)
+    # program_id is int32 and Triton does not promote int multiplies: at
+    # 1M-context prefill shapes the row base pid_b * s_out_b crosses 2^31
+    # (bisected: first IMA exactly one step past B*msl = 2^31), so every
+    # per-row base below is computed in int64. Same class of bug as the
+    # vLLM-side indexer overflow (vllm#50576).
+    pid_b = tl.program_id(0).to(tl.int64)
     pid_n = tl.program_id(1)
     offs_n = pid_n * BN + tl.arange(0, BN)
     n_mask = offs_n < PADDED_EFF
